requirements.txt
appdirs==1.4.4
argcomplete==3.6.3
customtkinter==5.2.2
darkdetect==0.8.0
escpos==2.0.0
future==1.0.0
numpy==2.3.5
packaging==25.0
pillow==12.0.0
PyAudio==0.2.14
pyserial==3.5
python-barcode==0.16.1
pyusb==1.3.1
PyYAML==6.0.3
qrcode==8.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
viivakoodi==0.8.0
